+++ R/tokenize.R
@@ -20,13 +20,37 @@
 #' @name tokenize
 NULL
 
-#' @export
 #' @rdname tokenize
-tokenize <-
-  function(text,
-           match_option = Match$ALL,
-           stopwords = TRUE) {
-    return(purrr::map(
+#' @export
+#' @importFrom dplyr bind_rows
+tokenize <- function(text,
+                     match_option = Match$ALL,
+                     stopwords = TRUE) {
+  dplyr::bind_rows(tokenize_raw(text, match_option, stopwords), .id = "sent")
+}
+
+#' @rdname tokenize
+#' @export
+tokenize_tbl <- tokenize
+
+#' @rdname tokenize
+#' @export
+#' @importFrom purrr map
+tokenize_tidytext <- function(text,
+                              match_option = Match$ALL,
+                              stopwords = TRUE) {
+  purrr::map(tokenize_raw(text, match_option, stopwords),
+             ~ paste0(.x$form, "/", .x$tag))
+}
+
+#' @rdname tokenize
+#' @export
+tokenize_tidy <- tokenize_tt <- tokenize_tidytext
+
+#' @importFrom purrr map_chr map_int
+tokenize_raw <- function(text, match_option, stopwords) {
+  purrr::map(
+    purrr::map(
       text,
       ~ analyze(
         text = .x,
@@ -34,47 +58,7 @@ tokenize <-
         match_option = match_option,
         stopwords = stopwords
       )[[1]][1]
-    ))
-  }
-
-
-#' @rdname tokenize
-#' @export
-#' @importFrom dplyr bind_rows
-tokenize_tibble <- function(text,
-                            match_option = Match$ALL,
-                            stopwords = FALSE) {
-  dplyr::bind_rows(tokenize_raw(text, match_option, stopwords), .id = "unique")
-}
-
-#' @rdname tokenize
-#' @export
-#' @importFrom purrr map
-tokenize_tidytext <- function(text,
-                              match_option = Match$ALL,
-                              stopwords = FALSE) {
-  purrr::map(tokenize_raw(text, match_option, stopwords),
-             ~ paste0(.x$form, "/", .x$tag))
-}
-
-
-#' @rdname tokenize
-#' @export
-tokenize_tbl <- tokenize_tibble
-
-#' @rdname tokenize
-#' @export
-tokenize_tt <- tokenize_tidytext
-
-#' @rdname tokenize
-#' @export
-tokenize_tidy <- tokenize_tidytext
-
-#' @importFrom purrr map_chr map_int
-tokenize_raw <- function(text, match_option, stopwords) {
-  res <- tokenize(text, match_option, stopwords)
-  purrr::map(
-    res,
+    ),
     ~ tibble::tibble(
       form = purrr::map_chr(.x$Token, ~ .x$form),
       tag = purrr::map_chr(.x$Token, ~ .x$tag),